+++ packages/acter_notifify/lib/util.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
 import 'package:acter_notifify/platform/android.dart';
 import 'package:acter_notifify/local.dart';
 import 'package:acter_notifify/platform/windows.dart';
@@ -11,7 +12,7 @@ final DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
 final useLocal = Platform.isAndroid ||
     Platform.isIOS ||
     Platform.isMacOS ||
-    Platform.isLinux; // || Platform.isMacOS;
+    Platform.isLinux;
 
 final usePush = Platform.isAndroid || Platform.isIOS;
 
@@ -80,4 +81,53 @@ Future<String> deviceName() async {
   } else {
     return '(unknown)';
   }
+}
+
+(String, String?) genTitleAndBody(NotificationItem notification) =>
+    switch (notification.pushStyle()) {
+      "comment" => _titleAndBodyForComment(notification),
+      "reaction" => _titleAndBodyForReaction(notification),
+      _ => _fallbackTitleAndBody(notification),
+    };
+
+(String, String?) _fallbackTitleAndBody(NotificationItem notification) =>
+    (notification.title(), notification.body()?.body());
+
+String _parentPart(NotificationItemParent parent) {
+  final emoji = parent.emoji();
+  final title = switch (parent.objectTypeStr()) {
+    'news' => "boost",
+    _ => parent.title(),
+  };
+  return "$emoji $title";
+}
+
+(String, String?) _titleAndBodyForComment(NotificationItem notification) {
+  final parent = notification.parent();
+  String title = "💬 Comment";
+  if (parent != null) {
+    final parentInfo = _parentPart(parent);
+    title = "$title on $parentInfo";
+  }
+
+  final comment = notification.body()?.body();
+  final sender = notification.sender();
+  final username = sender.displayName() ?? sender.userId();
+
+  return (title, "$username: $comment");
+}
+
+(String, String?) _titleAndBodyForReaction(NotificationItem notification) {
+  final parent = notification.parent();
+  final reaction = notification.reactionKey() ?? '❤️';
+  String title = '"$reaction"';
+  if (parent != null) {
+    final parentInfo = _parentPart(parent);
+    title = "$title to $parentInfo";
+  }
+
+  final sender = notification.sender();
+  final username = sender.displayName() ?? sender.userId();
+
+  return (title, username);
 }
